+++ src/common/config/main.php
@@ -40,7 +40,16 @@ $config = [
 if (YII_DEBUG && !YII_ENV_TEST) {
     // configuration adjustments for 'dev' environment
     $config['bootstrap'][] = 'debug';
-    $config['modules']['debug'] = 'yii\debug\Module';
+    $config['modules']['debug'] = [
+        'class' => 'yii\debug\Module',
+        'allowedIPs' => $params['debug_ips'],
+        'panels' => [
+            'hiresource' => [
+                'class' => 'hiqdev\hiart\DebugPanel',
+            ]
+        ]
+    ];
+
 
     $config['bootstrap'][] = 'gii';
     $config['modules']['gii'] = 'yii\gii\Module';